+++ node/utils/index.ts
@@ -34,8 +34,6 @@ export const currentSchema = (entity: string) => {
   })
 }
 
-export const rolesVbaseId = 'allRolesVbId'
-
 export const currentRoleNames = (locale: any) => {
   return roleNames[locale] ?? roleNames['en-US']
 }